Move translations to JS
