+++ app/assets/assets/help/en/contribute/translate.md
@@ -4,10 +4,10 @@ of replacing strings in files.
 #### If you're comfortable with git
 
 1. Fork the [repo](https://github.com/CaptainFact/captain-fact-frontend)
-2. Edit or create files in `app/assets/assets/locales/${YOUR_LOCALE}/${namespace}.json`
+2. Edit or create files in `app/i18n/${YOUR_LOCALE}/${namespace}.js`
    to translate the user interface
-3. Edit or create files in `app/assets/assets/help/${YOUR_LOCALE}/${page}.md` to translate help pages
-    like this one
+3. Edit or create files in `app/assets/assets/help/${YOUR_LOCALE}/${page}.md`
+   to translate help pages like this one
 4. Make a pull request against `staging` branch
 
 #### If you don't know what the heck git is about
