+++ app/assets/assets/help/fr/contribute/translate.md
@@ -4,7 +4,7 @@ des chaînes de caractères dans des fichiers.
 #### Si vous êtes à l'aise avec git
 
 1. Forker le [dépôt](https://github.com/CaptainFact/captain-fact-frontend)
-2. Editer ou créer les fichiers dans `app/assets/assets/locales/${YOUR_LOCALE}/${namespace}.json`
+2. Editer ou créer les fichiers dans `app/i18n/${YOUR_LOCALE}/${namespace}.js`
    pour traduire l'interface utilisateur
 3. Editer ou créer les fichiers dans `app/assets/assets/help/${YOUR_LOCALE}/${page}.md`
    pour traduire les pages d'aide (comme celle-ci).
