+++ app/i18n/en/home.js
@@ -1,4 +1,4 @@
-{
+export default {
   "presentation1": "Internet is filled with rich and abundant data. Information travels fast and we don't have the time to check that everything we see and hear is correct.",
   "presentation2": "is all about people debating and helping each other to verify that data by sourcing it, voting and chatting about facts.",
   "journeyStart": "You can start your journey by checking the public {{videoLink}}",
