+++ app/i18n/en/index.js
@@ -1,14 +1,12 @@
-// TODO Declare translations directly in JS
-
-import achievements from '../../assets/assets/locales/en/achievements.json'
-import errors from '../../assets/assets/locales/en/errors.json'
-import help from '../../assets/assets/locales/en/help.json'
-import history from '../../assets/assets/locales/en/history.json'
-import home from '../../assets/assets/locales/en/home.json'
-import main from '../../assets/assets/locales/en/main.json'
-import user from '../../assets/assets/locales/en/user.json'
-import videoDebate from '../../assets/assets/locales/en/videoDebate.json'
-import moderation from '../../assets/assets/locales/en/moderation.json'
+import achievements from './achievements'
+import errors from './errors'
+import help from './help'
+import history from './history'
+import home from './home'
+import main from './main'
+import user from './user'
+import videoDebate from './videoDebate'
+import moderation from './moderation'
 
 export default {
   achievements,
