+++ app/i18n/en/moderation.js
@@ -1,4 +1,4 @@
-{
+export default {
   "header": {
     "date": "Date"
   },
