+++ app/i18n/fr/history.js
@@ -1,4 +1,4 @@
-{
+export default {
   "compare_show": "Comparer",
   "compare_hide": "Masquer",
   "compareAll": "Tout comparer",
