+++ app/i18n/fr/home.js
@@ -1,4 +1,4 @@
-{
+export default {
   "presentation1" : "Internet est rempli de données riches et diverses. L'information voyage vite et nous n'avons pas le temps de vérifier si tout ce que nous entendons et voyons est correct.",
   "presentation2" : "nous permet de débattre et de nous entraider pour vérifier ces données en les sourçant, en votant et en discutant autour de faits.",
   "journeyStart": "Vous pouvez démarrer votre aventure en allant voir les {{videoLink}} publiques",
