+++ app/i18n/fr/index.js
@@ -1,14 +1,12 @@
-// TODO Declare translations directly in JS
-
-import achievements from '../../assets/assets/locales/fr/achievements.json'
-import errors from '../../assets/assets/locales/fr/errors.json'
-import help from '../../assets/assets/locales/fr/help.json'
-import history from '../../assets/assets/locales/fr/history.json'
-import home from '../../assets/assets/locales/fr/home.json'
-import main from '../../assets/assets/locales/fr/main.json'
-import user from '../../assets/assets/locales/fr/user.json'
-import videoDebate from '../../assets/assets/locales/fr/videoDebate.json'
-import moderation from '../../assets/assets/locales/fr/moderation.json'
+import achievements from './achievements'
+import errors from './errors'
+import help from './help'
+import history from './history'
+import home from './home'
+import main from './main'
+import user from './user'
+import videoDebate from './videoDebate'
+import moderation from './moderation'
 
 export default {
   achievements,
