+++ app/i18n/fr/moderation.js
@@ -1,4 +1,4 @@
-{
+export default {
   "header": {
     "date": "Date"
   },
